FIX: Tests used too small objects
We do not account for very small free blocks in swap as this would be more
memory consuming than just reserving them. Consequently, the order of
allocation matters for big objects. We could have a feature on the list
that is called 'defraction' to prevent this, but this is time consuming
anyways